Add text style custom modifier and other fixes
