+++ DeUrgenta/Features/EmergencyBackpack/View/MyBackpackView.swift
@@ -1,14 +1,7 @@
 import SwiftUI
 
 struct MyBackpackView: View {
-    let items = [
-        CategoryItem(title: AppStrings.CategoryItem.foodTitle.localized(), icon: "icon-syrup", type: .food),
-        CategoryItem(title: AppStrings.CategoryItem.hygieneItemsTitle.localized(), icon: "icon-cotton-swabs", type: .hygieneItems),
-        CategoryItem(title: AppStrings.CategoryItem.firstAidKitTitle.localized(), icon: "icon-capsules", type: .firstAidKit),
-        CategoryItem(title: AppStrings.CategoryItem.docsTitle.localized(), icon: "icon-prescription", type: .docs),
-        CategoryItem(title: AppStrings.CategoryItem.survivalItemsTitle.localized(), icon: "icon-medical-mask", type: .survivalItems),
-        CategoryItem(title: AppStrings.CategoryItem.mixedTitle.localized(), icon: "icon-pharmacy", type: .mixed),
-    ]
+   @StateObject var viewModel = EmergencyBackpackViewModel()
 
     var body: some View {
         ScrollView {
@@ -20,12 +13,10 @@ struct MyBackpackView: View {
 
                 VStack {
                     Text(AppStrings.MyBackpackView.firstParagraph.localized())
-                        .font(.custom("IBMPlexSans-Regular", size: 16))
-                        .foregroundColor(.darkText)
+                        .modifier(PrimaryTextStyle())
 
                     Text(AppStrings.MyBackpackView.secondParagraph.localized())
-                        .font(.custom("IBMPlexSans-Regular", size: 16))
-                        .foregroundColor(.darkText)
+                        .modifier(PrimaryTextStyle())
                         .padding(.top, 15)
                 }
                 .lineSpacing(10)
@@ -61,10 +52,11 @@ struct MyBackpackView: View {
 
                 Text(AppStrings.MyBackpackView.backpackitemsHeader.localized())
                     .font(.custom("IBMPlexSans-SemiBold", size: 20))
+                    .foregroundColor(.darkText)
                     .padding(.vertical)
 
                 LazyVStack {
-                    ForEach(items, id: \.id) { item in
+                    ForEach(viewModel.items, id: \.id) { item in
                         BackpackItemView(item: item)
                     }
                 }
